Move report_table, report_settings_form, report_display_options to classes folder
