+++ classes/report_settings_form.php
@@ -15,36 +15,26 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * This file defines the options for the quiz answersheets report.
+ * This file defines the setting form for the quiz answersheets report.
  *
  * @package   quiz_answersheets
  * @copyright 2019 The Open University
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+namespace quiz_answersheets;
 
 defined('MOODLE_INTERNAL') || die();
 
-require_once($CFG->dirroot . '/mod/quiz/report/attemptsreport_options.php');
+require_once($CFG->dirroot . '/mod/quiz/report/attemptsreport_form.php');
 
 /**
- * This file defines the options for the quiz answersheets report.
+ * This file defines the setting form for the quiz answersheets report.
  *
  * @package   quiz_answersheets
  * @copyright 2019 The Open University
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class quiz_answersheets_options extends mod_quiz_attempts_report_options {
-    public function __construct($mode, $quiz, $cm, $course) {
-        parent::__construct($mode, $quiz, $cm, $course);
-        $this->attempts = quiz_attempts_report::ENROLLED_ALL;
-    }
-
-    public function resolve_dependencies() {
-        parent::resolve_dependencies();
-        // We only want to show the checkbox to delete attempts
-        // if the user has permissions and if the report mode is showing attempts.
-        $this->checkboxcolumn = has_capability('mod/quiz:deleteattempts', context_module::instance($this->cm->id))
-                && ($this->attempts != quiz_attempts_report::ENROLLED_WITHOUT);
-    }
+class report_settings_form extends \mod_quiz_attempts_report_form {
 
 }
